Revert code_editor fixture addition and use a previously defined fixture
